+++ packages/dht/src/connection/IConnectionSource.ts
@@ -4,7 +4,7 @@ export enum Event {
     CONNECTED = 'streamr:dht:connectionsource:connected'
 }
 
-export interface ConnectionSource {
+export interface IConnectionSource {
     on(event: Event.CONNECTED, listener: (connection: Connection) => void): this
     once(event: Event.CONNECTED, listener: (connection: Connection) => void): this
 }
