+++ packages/dht/src/connection/WebSocket/WebSocketConnector.ts
@@ -1,9 +1,9 @@
 import { EventEmitter } from 'events'
-import { ConnectionSource, Event as ConnectionSourceEvent } from '../ConnectionSource'
+import { IConnectionSource, Event as ConnectionSourceEvent } from '../IConnectionSource'
 import { ClientWebSocket } from './ClientWebSocket'
 import { Event as ConnectionEvent, Connection } from '../Connection'
 
-export class WebSocketConnector extends EventEmitter implements ConnectionSource {
+export class WebSocketConnector extends EventEmitter implements IConnectionSource {
 
     connect({ host, port, url }: { host?: string; port?: number; url?: string; } = {}): ClientWebSocket {
         const socket = new ClientWebSocket()
