+++ packages/dht/test/integration/WebSocket.test.ts
@@ -2,7 +2,7 @@
 
 import { WebSocketConnector } from "../../src/connection/WebSocket/WebSocketConnector"
 import { WebSocketServer } from "../../src/connection/WebSocket/WebSocketServer"
-import { Event as ConnectionSourceEvent } from '../../src/connection/ConnectionSource'
+import { Event as ConnectionSourceEvent } from '../../src/connection/IConnectionSource'
 import { Connection, Event as ConnectionEvent } from "../../src/connection/Connection"
 
 describe('WebSocket', () => {
